backend/pyproject.toml
[project]
name = "anonymous-hf"
version = "0.1.0"
description = "Anonymous HuggingFace - proxy service for double-blind peer review"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.2",
    "django-cors-headers>=4.9.0",
    "django-q2>=1.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "gunicorn>=25.0.3",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "requests>=2.32.5",
    "whitenoise>=6.11.0",
]

[project.optional-dependencies]
dev = []

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "anonymous_hf.settings"
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "responses>=0.25.8",
    "ruff>=0.15.0",
]
